+++ src/risedevtool/src/task/dummy_service.rs
@@ -1,0 +1,42 @@
+// Copyright 2024 RisingWave Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use crate::{ExecuteContext, Task};
+
+pub struct DummyService {
+    id: String,
+}
+
+impl DummyService {
+    pub fn new(id: &str) -> Self {
+        Self { id: id.to_string() }
+    }
+}
+
+impl Task for DummyService {
+    fn execute(&mut self, ctx: &mut ExecuteContext<impl std::io::Write>) -> anyhow::Result<()> {
+        ctx.service(self);
+        writeln!(
+            &mut ctx.log,
+            "{} is a dummy service. Please ensure it's correctly configured on your own! 🙏",
+            self.id
+        )?;
+        ctx.complete_spin();
+        Ok(())
+    }
+
+    fn id(&self) -> String {
+        self.id.clone()
+    }
+}
